# docs/references/api/swagger.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is the API specification of the Epinio API",
    "title": "Epinio API",
    "version": "1.0.0"
  },
  "basePath": "/api/v1",
  "paths": {
...
    "/namespaces/{Namespace}/applications/{App}/environment": {
      "get": {
        "tags": [
          "app-env"
        ],
        "summary": "Return the environment variable assignments for the `App` in the namespace`.",
        "operationId": "EnvList",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "App",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/EnvListResponse"
          }
        }
      },
      "post": {
        "tags": [
          "app-env"
        ],
        "summary": "Create/modify the posted environment variable assignments for the `App` in the `Namespace`.",
        "operationId": "EnvSet",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "App",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/EnvVariableMap"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/EnvSetResponse"
          }
        }
      }
    },
...
    "/namespaces/{Namespace}/configurations": {
      "get": {
        "tags": [
          "configuration"
        ],
        "summary": "Return list of configurations in the `Namespace`.",
        "operationId": "Configurations",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ConfigurationsResponse"
          }
        }
      },
      "post": {
        "tags": [
          "configuration"
        ],
        "summary": "Create the posted new configuration in the `Namespace`.",
        "operationId": "ConfigurationCreate",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "Configuration",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ConfigurationCreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ConfigurationCreateResponse"
          }
        }
      },
      "delete": {
        "summary": "Delete the named `Configurations` in the `Namespace`.",
        "operationId": "ConfigurationBatchDelete",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "in: url",
            "name": "Configurations",
            "in": "query"
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ConfigurationDeleteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ConfigurationDeleteResponse"
          }
        }
      }
    },
...
    "/namespaces/{Namespace}/services/{Service}/unbind": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Unbind the named `Service` in the `Namespace` from an App.",
        "operationId": "ServiceUnbind",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "Service",
            "in": "path",
            "required": true
          },
          {
            "name": "Configuration",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ServiceUnbindRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ServiceUnbindResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ConfigurationCreateRequest": {
      "description": "ConfigurationCreateRequest represents and contains the data needed to\ncreate a configuration instance",
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Data"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ConfigurationDeleteRequest": {
      "description": "ConfigurationDeleteRequest represents and contains the data needed to delete a configuration",
      "type": "object",
      "properties": {
        "unbind": {
          "type": "boolean",
          "x-go-name": "Unbind"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ConfigurationDeleteResponse": {
      "description": "ConfigurationDeleteResponse represents the server's response to a successful configuration deletion",
      "type": "object",
      "properties": {
        "boundapps": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "BoundApps"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "ConfigurationRef": {
      "description": "ConfigurationRef references a Configuration by name and namespace",
      "type": "object",
      "properties": {
        "createdAt": {
          "$ref": "#/definitions/Time"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "ConfigurationResponse": {
      "description": "ConfigurationResponse represents the data of a single configuration instance",
      "type": "object",
      "properties": {
        "configuration": {
          "$ref": "#/definitions/ConfigurationShowResponse"
        },
        "meta": {
          "$ref": "#/definitions/ConfigurationRef"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ConfigurationResponseList": {
      "description": "ConfigurationResponseList represents a collection of configuration instance",
      "type": "array",
      "items": {
        "$ref": "#/definitions/ConfigurationResponse"
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ConfigurationShowResponse": {
      "description": "ConfigurationShowResponse contains details about a configuration",
      "type": "object",
      "properties": {
        "boundapps": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "BoundApps"
        },
        "details": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "origin": {
          "type": "string",
          "x-go-name": "Origin"
        },
        "siblings": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Siblings"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        },
        "user": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "EnvVariableMap": {
      "description": "EnvVariableMap is a collection of EVs as a map. It is used for Set Requests, and as\nList Responses",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "Response": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "ServiceUnbindRequest": {
      "type": "object",
      "properties": {
        "app_name": {
          "type": "string",
          "x-go-name": "AppName"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "Time": {
      "description": "+protobuf.options.marshal=false\n+protobuf.as=Timestamp\n+protobuf.options.(gogoproto.goproto_stringer)=false",
      "type": "object",
      "title": "Time is a wrapper around time.Time which supports correct\nmarshaling to YAML and JSON.  Wrappers are provided for many\nof the factory methods that the time package offers.",
      "x-go-package": "k8s.io/apimachinery/pkg/apis/meta/v1"
    },
...
  },
  "responses": {
...
    "ConfigurationCreateResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Response"
      }
    },
    "ConfigurationDeleteResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ConfigurationDeleteResponse"
      }
    },
...
    "ConfigurationsResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ConfigurationResponseList"
      }
    },
    "EnvListResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/EnvVariableMap"
      }
    },
...
    "EnvSetResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Response"
      }
    },
...
    "ServiceUnbindResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Response"
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  },
  "security": [
    {
      "basicAuth": [
        "[]"
      ]
    }
  ]
}